% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{air_minum_limbah}
\alias{air_minum_limbah}
\title{Jumlah Pelanggan Air Minum dan Air Limbah PDAM Tirtawening}
\format{Tibble dengan 24 baris dan 4 variabel:
\describe{
  \item{\code{bulan}}{ }
  \item{\code{tahun}}{ }
  \item{\code{jumlah_pelanggan_air_minum}}{jumlah pelanggan air minum}
  \item{\code{jumlah_pelanggan_air_limbah}}{jumlah pelanggan air limbah}
}}
\source{
\url{http://data.bandung.go.id/dataset/jumlah-pelanggan-air-minum-dan-air-limbah-pdam-tirtawening}
}
\usage{
air_minum_limbah
}
\description{
Dataset ini berisi mengenai jumlah pelanggan air minum dan air limbah PDAM Tirtawening Kota Bandung.
}
\keyword{datasets}
