% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{anak_sungai}
\alias{anak_sungai}
\title{Panjang Anak Sungai Citarum di Kota Bandung}
\format{Tibble dengan 15 baris dan 2 variabel:
\describe{
  \item{\code{anak_sungai}}{nama anak sungai}
  \item{\code{Panjang}}{panjang anak sungai (km)}
}}
\source{
\url{http://data.bandung.go.id/dataset/panjang-anak-sungai-citarum-di-kota-bandung}
}
\usage{
anak_sungai
}
\description{
Dataset ini berisi data panjang anak sungai citarum yang terdiri dari 15 anak sungai.
}
\keyword{datasets}
