% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{angin}
\alias{angin}
\title{Kecepatan Angin Menurut Bulan Di Kota Bandung}
\format{Tibble dengan 60 baris dan 3 variabel:
\describe{
  \item{\code{tanggal}}{tanggal pengukuran}
  \item{\code{kec_ratarata}}{kecepatan angin rata-rata (knot)}
  \item{\code{kec_terbesar}}{kecepatan angin terbesar (knot)}
}}
\source{
\url{http://data.bandung.go.id/dataset/kecepatan-angin-menurut-bulan-di-kota-bandung-2015}
}
\usage{
angin
}
\description{
Dataset ini berisi Kecepatan Angin Menurut Bulan Di Kota Bandung.
}
\keyword{datasets}
