% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{bak_sampah}
\alias{bak_sampah}
\title{Kapasitas Bak Penampungan Sampah Untuk Wilayah Kota Bandung}
\format{Tibble dengan 20 baris dan 2 variabel:
\describe{
  \item{\code{nama_bak}}{nama bak sampah}
  \item{\code{kapasitas}}{kapasitas (m3)}
}}
\source{
\url{http://data.bandung.go.id/dataset/kapasitas-bak-penampungan-sampah-untuk-wilayah-kota-bandung}
}
\usage{
bak_sampah
}
\description{
Dataset kapasitas bak penampungan sampah untuk wilayah kota Bandung
}
\keyword{datasets}
