% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{bank_sampah}
\alias{bank_sampah}
\title{Unit Bank Sampah Binaan Bank Sampah Resik - Tahun 2018}
\format{Tibble dengan 110 baris dan 5 variabel:
\describe{
  \item{\code{Kategori}}{kategori instansi}
  \item{\code{Instansi}}{nama instansi}
  \item{\code{Alamat}}{alamat instansi}
  \item{\code{Kelurahan}}{lokasi kelurahan instansi}
  \item{\code{Kecamatan}}{lokasi kecamatan instansi}
}}
\source{
\url{http://data.bandung.go.id/dataset/unit-bank-sampah-binaan-bank-sampah-resik}
}
\usage{
bank_sampah
}
\description{
Dataset ini data unit bank sampah binaan bank sampah resik PD Kebersihan Kota Bandung.
}
\keyword{datasets}
