% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{bendungan}
\alias{bendungan}
\title{Bendungan di Kota Bandung}
\format{Tibble dengan 42 baris dan 6 variabel:
\describe{
  \item{\code{No}}{ }
  \item{\code{Bendungan}}{nama bendungan}
  \item{\code{Sumber_air}}{sumber air bendungan}
  \item{\code{Kelurahan}}{lokasi kelurahan bendungan}
  \item{\code{Kecamatan}}{lokasi kecamatan bendungan}
  \item{\code{Keterangan}}{keterangan terkait pemanfaatan bendungan}
}}
\source{
\url{http://data.bandung.go.id/dataset/bendungan-di-kota-bandung}
}
\usage{
bendungan
}
\description{
Dataset ini berisi data bendungan di Kota Bandung
}
\keyword{datasets}
