% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{cakupan_am}
\alias{cakupan_am}
\title{Rekapitulasi Jumlah Cakupan Air Minum Di Tingkat RT (Rukun Tetangga)}
\format{Tibble dengan 120 baris dan 6 variabel:
\describe{
  \item{\code{No.}}{ }
  \item{\code{Kecamatan}}{nama kecamatan}
  \item{\code{Kelurahan}}{nama kelurahan}
  \item{\code{Jumlah_krt}}{jumlah kepala rumah tangga}
  \item{\code{Jumlah_kk}}{jumlah kepala keluarga}
  \item{\code{Jumlah_masy_terlayani}}{jumlah masyarakat terlayani Sarana Air Minum untuk minum, mandi, dan cuci (perpipaan atau non perpipaan terlindungi yang layak) (rumah tangga)}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
cakupan_am
}
\description{
Dataset ini berisi data rekapitulasi jumlah cakupan air minum di tingkat RT (Rukun Tetangga).
}
\keyword{datasets}
