% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{cakupan_am_rt}
\alias{cakupan_am_rt}
\title{Rekapitulasi Jumlah Cakupan Air Minum Di Tingkat RT (Rukun Tetangga)}
\format{Tibble dengan 2681 baris dan 9 variabel:
\describe{
  \item{\code{No.}}{ }
  \item{\code{Kecamatan}}{nama kecamatan}
  \item{\code{Kelurahan}}{nama kelurahan}
  \item{\code{RW}}{rukun warga}
  \item{\code{RT}}{rukun tetangga}
  \item{\code{Jumlah_krt}}{jumlah kepala rumah tangga}
  \item{\code{Jumlah_kk}}{jumlah kepala keluarga}
  \item{\code{Jumlah_masy_terlayani}}{jumlah masyarakat terlayani Sarana Air Minum untuk minum, mandi, dan cuci (perpipaan atau non perpipaan terlindungi yang layak) (rumah tangga)}
  \item{\code{Persen_terlayani}}{prosentase masyarakat terlayani Sarana Air Minum untuk minum, mandi, dan cuci (perpipaan atau non perpipaan terlindungi yang layak)}
}}
\source{
\url{http://data.bandung.go.id/dataset/rekapitulasi-jumlah-cakupan-air-minum-di-tingkat-rt-rukun-tetangga}
}
\usage{
cakupan_am_rt
}
\description{
Dataset ini berisi data rekapitulasi jumlah cakupan air minum di tingkat RT (Rukun Tetangga).
}
\keyword{datasets}
