% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{cuaca}
\alias{cuaca}
\title{Cuaca dan Curah Hujan Kota Bandung}
\format{Tibble dengan 26 baris dan 9 variabel:
\describe{
  \item{\code{bulan}}{ }
  \item{\code{temperatur_ratarata}}{temperatur rata-rata (oC)}
  \item{\code{temperatur_max}}{temperatur maksimum (oC)}
  \item{\code{temperatur_min}}{temperatur minimum (oC)}
  \item{\code{curah_hujan}}{curah hujan (mm)}
  \item{\code{hari_hujan}}{hari hujan (hari)}
  \item{\code{lpm}}{lpm persen}
}}
\source{
\url{http://data.bandung.go.id/dataset/cuaca-dan-curah-hujan-kota-bandung-tahun-2013}
}
\usage{
cuaca
}
\description{
Dataset mengenai cuaca dan curah hujan di kota Bandung
}
\keyword{datasets}
