% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{izin_lingkungan}
\alias{izin_lingkungan}
\title{Izin Lingkungan di Kota Bandung}
\format{Tibble dengan 321 baris dan 7 variabel:
\describe{
  \item{\code{no}}{}
  \item{\code{nomor_izin}}{}
  \item{\code{tahun}}{}
  \item{\code{jenis_lingkungan}}{}
  \item{\code{nama_pemohon}}{}
  \item{\code{nama_perusahaan}}{}
  \item{\code{bentuk_perusahaan}}{}
}}
\source{
\url{http://data.bandung.go.id/dataset/izin-lingkungan-di-kota-bandung}
}
\usage{
izin_lingkungan
}
\description{
Dataset ini berisi data izin lingkungan di Kota Bandung.
}
\keyword{datasets}
