% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{KangPisMan}
\alias{KangPisMan}
\title{Kegiatan Sosialisasi KangPisMan Kota Bandung}
\format{Tibble dengan 27 baris dan 5 variabel:
\describe{
  \item{\code{No}}{ }
  \item{\code{Tanggal}}{tanggal pelaksanaan sosialisasi}
  \item{\code{Kegiatan}}{nama kegiatan}
  \item{\code{Narasumber}}{narasumber kegiatan sosialisasi}
  \item{\code{Tempat}}{lokasi sosialisasi}
}}
\source{
\url{http://data.bandung.go.id/dataset/kegiatan-sosialisasi-kangpisman-kota-bandung}
}
\usage{
KangPisMan
}
\description{
Dataset ini berisi data kegiatan sosialisasi KangPisMan Kota Bandung dari tahun 2018 sampai 2019.
}
\keyword{datasets}
