% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{keterangan_tpa}
\alias{keterangan_tpa}
\title{Data Keterangan TPA PD Kebersihan Tahun 2005 - 2015}
\format{Tibble dengan 6 baris dan 3 variabel:
\describe{
  \item{\code{no}}{}
  \item{\code{nama_tpa}}{}
  \item{\code{keterangan}}{}
}}
\source{
\url{http://data.bandung.go.id/dataset/data-keterangan-tpa-pd-kebersihan-tahun-2005-2015}
}
\usage{
keterangan_tpa
}
\description{
Dataset ini berisi mengenai data keterangan Tempat Pembuangan Akhir (TPA) Perusahaan Daerah Kebersihan Kota Bandung Tahun 2005-2015.
}
\keyword{datasets}
