% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{kual_sungai}
\alias{kual_sungai}
\title{Kualitas Air Sungai Kota Bandung Tahun 2018}
\format{Tibble dengan 56 baris dan 5 variabel:
\describe{
  \item{\code{No}}{ }
  \item{\code{Sungai}}{nama sungai}
  \item{\code{Storet_Score}}{nilai indeks STORET}
  \item{\code{Golongan}}{golongan kelas sungai berdasarkan nilai indeks STORET}
  \item{\code{Kategori}}{kategori sungai berdasarkan golongan kelas sungai}
}}
\source{
\url{http://data.bandung.go.id/dataset/kualitas-air-sungai-kota-bandung}
}
\usage{
kual_sungai
}
\description{
Dataset ini berisi data kualitas air sungai di Kota Bandung.
}
\keyword{datasets}
