% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{mata_air}
\alias{mata_air}
\title{Sumber Mata Air di Kota Bandung}
\format{Tibble dengan 82 baris 9 variabel:
\describe{
  \item{\code{No}}{ }
  \item{\code{Wilayah}}{nama wilayah}
  \item{\code{Mata_air}}{nama mata air}
  \item{\code{Saluran}}{saluran/sungai lokasi air mata air mengalir}
  \item{\code{Kampung}}{nama kampung mata air}
  \item{\code{Kelurahan}}{nama kelurahan}
  \item{\code{Kecamatan}}{nama kecamatan}
  \item{\code{Debit}}{debit rata-rata (l/detik)}
  \item{\code{Penggunaan}}{pemanfaatan mata air}
  \item{\code{Keterangan}}{keterangan kepemilikan lahan mata air}
}}
\source{
\url{http://data.bandung.go.id/dataset/sumber-mata-air-di-kota-bandung}
}
\usage{
mata_air
}
\description{
Dataset ini berisi data sumber mata air di Kota Bandung.
}
\keyword{datasets}
