% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{mitra_KangPisMan}
\alias{mitra_KangPisMan}
\title{Data Mitra Komersial KangPisman - Tahun 2019}
\format{Tibble dengan 27 baris dan 2 variabel:
\describe{
  \item{\code{Kategori}}{kategori kegiatan komersial}
  \item{\code{Mitra}}{nama mitra}
}}
\source{
\url{http://data.bandung.go.id/dataset/mitra-komersial-kangpisman}
}
\usage{
mitra_KangPisMan
}
\description{
Dataset ini berisi data mitra komersial KangPisMan.
}
\keyword{datasets}
