% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{pelanggan_am}
\alias{pelanggan_am}
\title{Jumlah Pelanggan Air Minum Menurut Golongan Tarif di Kota Bandung}
\format{Tibble dengan 396 baris dan 4 variabel:
\describe{
  \item{\code{Golongan_tarif}}{golongan tarif}
  \item{\code{Bulan}}{bulan pencatatan}
  \item{\code{Jumlah_pelanggan}}{jumlah pelanggan}
  \item{\code{Tahun}}{tahun pencatatan}
}}
\source{
\url{http://data.bandung.go.id/dataset/jumlah-pelanggan-air-minum-menurut-golongan-tarif-di-kota-bandung}
}
\usage{
pelanggan_am
}
\description{
Dataset ini berisi mengenai data jumlah pelanggan air minum menurut golongan tarif di Kota Bandung.
}
\note{
Golongan tarif : 1A (Sosial Umum), 1B (Sosial Khusus), 2A1 (Rumah Tangga), 2A2 (Rumah Tangga), 2A3 (Rumah Tangga), 2A4 (Rumah Tangga), 2B (Instansi), 3A (Niaga Kecil), 3B (Niaga Menengah/Besar), 4A (Industri Kecil), 4B (Industri Menengah/Besar)
}
\keyword{datasets}
