% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{pengangkutan_tps}
\alias{pengangkutan_tps}
\title{Data Pengangkutan TPS Tahun 2005-2015}
\format{Tibble dengan 11 baris dan 9 variabel:
\describe{
  \item{\code{tahun}}{tahun pencatatan}
  \item{\code{gerobak}}{jumlah gerobak sampah}
  \item{\code{truk}}{jumlah truk sampah}
  \item{\code{kendaraan_kecil_pick_up}}{jumlah kendaraan kecil pick up}
  \item{\code{motor_sampah}}{jumlah motor sampah}
  \item{\code{container}}{jumlah kontainer sampah}
  \item{\code{tps}}{jumlah tps}
  \item{\code{tps_3r}}{jumlah tps3r}
  \item{\code{alat_berat}}{jumlah alat berat}
}}
\source{
\url{http://data.bandung.go.id/dataset/data-pengangkutan-tps-tahun-2005-2015/resource/7c755a73-30a7-4543-8f9d-ad21c30c349a}
}
\usage{
pengangkutan_tps
}
\description{
Dataset ini berisi mengenai jumlah moda pengangkut sampah dan jumlah Tempat Pembuangan Sampah (TPS) Tahun 2005-2015.
}
\keyword{datasets}
