% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{penggunaan_tanah}
\alias{penggunaan_tanah}
\title{Penggunaan Tanah Berdasarkan Jenis Penggunaan di Kota Bandung}
\format{Tibble dengan 32 baris dan 2 variabel:
\describe{
  \item{\code{jenis_penggunaan_tanah}}{jenis penggunaan tanah}
  \item{\code{luas_ha}}{luas tanah (Ha)}
}}
\source{
\url{http://data.bandung.go.id/dataset/penggunaan-tanah-berdasarkan-jenis-penggunaan-di-kota-bandung}
}
\usage{
penggunaan_tanah
}
\description{
Dataset ini berisi data penggunaan tanah berdasarkan jenis penggunaan di Kota Bandung.
}
\keyword{datasets}
