% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{pengomposan}
\alias{pengomposan}
\title{Data Pengomposan dan Pencacahan di Bidang Pengelolaan TPA}
\format{Tibble dengan 14 baris dan 8 variabel:
\describe{
  \item{\code{no}}{}
  \item{\code{tps}}{nama tps}
  \item{\code{tahun}}{tahun pendataan}
  \item{\code{sampah_organik}}{volume sampah organik (m3)}
  \item{\code{sampah_anorganik}}{volume sampah anorganik (m3)}
  \item{\code{sampah_cacahan}}{volume sampah cacahan (m3)}
  \item{\code{pengomposan}}{volume pengomposan (m3)}
  \item{\code{produksi_kompos}}{volume produksi kompos (m3)}
}}
\source{
\url{http://data.bandung.go.id/dataset/data-pengomposan-dan-pencacahan-di-bidang-pengelolaan-tpa}
}
\usage{
pengomposan
}
\description{
Data Pengomposan dan Pencacahan di Bidang Pengelolaan TPA
}
\keyword{datasets}
