% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sanitasi}
\alias{sanitasi}
\title{Rekapitulasi Jumlah Kepala Keluarga Yang Memenuhi Sanitasi Di Tingkat RT}
\format{Tibble dengan 2400 baris dan 10 variabel:
\describe{
  \item{\code{No.}}{ }
  \item{\code{Kecamatan}}{nama kecamatan}
  \item{\code{Kelurahan}}{nama kelurahan}
  \item{\code{RW}}{rukun warga}
  \item{\code{RT}}{rukun tetangga}
  \item{\code{Bkm_Status_Berdaya}}{status bkm}
  \item{\code{Jumlah_krt}}{jumlah kepala rumah tangga}
  \item{\code{Jumlah_kk}}{jumlah kepala keluarga}
  \item{\code{Jumlah_akses_mck}}{jumlah masyarakat memiliki akses MCK keluarga / MCK bersama (rumah tangga)}
  \item{\code{Jumlah_mck_sesuai}}{jumlah MCK keluarga/MCK bersama sesuai persyaratan teknis (memiliki kloset leher angsa yang terhubung dengan septic-tank) (rumah tangga)}
}}
\source{
\url{http://data.bandung.go.id/dataset/rekapitulasi-jumlah-kepala-keluarga-yang-memenuhi-sanitasi-di-tingkat-rt/resource/c9f5a1ff-fe25-4026-a01e-d3bb8403d632}
}
\usage{
sanitasi
}
\description{
Dataset ini berisi data rekapitulasi jumlah kepala keluarga yang memenuhi sanitasi di tingkat RT (Rukun Tetangga).
}
\keyword{datasets}
