% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sapuan_jalan}
\alias{sapuan_jalan}
\title{Data Hasil Sapuan Bulan Februari 2015 sampai Februari 2016}
\format{Tibble dengan 12 baris dan 8 variabel:
\describe{
  \item{\code{no}}{ }
  \item{\code{laporan}}{}
  \item{\code{jumlah_penyapu}}{}
  \item{\code{jumlah_hari_kerja_hari}}{}
  \item{\code{jumlah_kantong_trashbag}}{}
  \item{\code{jumlah_sampah_liter}}{}
  \item{\code{jumlah_sampah_m3}}{}
  \item{\code{jumlah_sampah_ton}}{}
}}
\source{
\url{http://data.bandung.go.id/dataset/data-hasil-sapuan-bulan-februari-2015-sampai-februari-2016}
}
\usage{
sapuan_jalan
}
\description{
Dataset ini berisi mengenai data hasil sapuan yang dilaksanakan oleh pihak ketiga mulai dari bulan Februari 2015 sampai dengan bulan Februari 2016.
}
\keyword{datasets}
