% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sumur_resapan}
\alias{sumur_resapan}
\title{Pembuatan Sumur Resapan Dan Lubang Biopori di Kota Bandung}
\format{Tibble dengan 30 baris dan 3 variabel:
\describe{
  \item{\code{nama_kecamatan}}{nama kecamatan}
  \item{\code{sumur_resapan}}{jumlah sumur resapan}
  \item{\code{lubang_biopori}}{jumlah lubang biopori}
}}
\source{
\url{http://data.bandung.go.id/dataset/pembuatan-sumur-resapan-dan-lubang-biopori-di-kota-bandung}
}
\usage{
sumur_resapan
}
\description{
Dataset ini berisi Rekapitulasi Pembuatan Sumur Resapan Dan Lubang Biopori di Kota Bandung.
}
\keyword{datasets}
