% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sw_jenis}
\alias{sw_jenis}
\title{Produksi Jenis Sampah Menurut Jenisnya di Kota Bandung}
\format{Tibble dengan 27 baris dan 4 variabel:
\describe{
  \item{\code{Tahun}}{  }
  \item{\code{Jenis_sampah}}{jenis sampah}
  \item{\code{Prod_sampah}}{produksi sampah (m3/hari)}
  \item{\code{Persentase}}{persentase tiap jenis sampah}
}}
\source{
\url{http://data.bandung.go.id/dataset/produksi-jenis-sampah-menurut-jenisnya-tahun-2014}
}
\usage{
sw_jenis
}
\description{
Dataset ini berisi produksi jenis sampah menurut jenisnya di Kota Bandung
}
\keyword{datasets}
