% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{sw_prod_sumber}
\alias{sw_prod_sumber}
\title{Rata-rata Produksi Sampah Berdasarkan Sumber Sampah di Kota Bandung}
\format{Tibble dengan 6 baris dan 2 variabel:
\describe{
  \item{\code{Sumber}}{sumber sampah}
  \item{\code{Prod_sampah}}{produksi sampah (ton)}
}}
\source{
\url{http://data.bandung.go.id/dataset/rata-rata-produksi-sampah-berdasarkan-sumber-sampah-di-kota-bandung}
}
\usage{
sw_prod_sumber
}
\description{
Dataset ini berisi data rata-rata produksi sampah berdasarkan sumber sampah di Kota Bandung.
}
\keyword{datasets}
