% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{taman}
\alias{taman}
\title{Luas Taman Kota Di Sub Wilayah Kota (SWK) Bandung Tahun 2018}
\format{Tibble dengan 6 baris dan 3 variabel:
\describe{
  \item{\code{Wilayah}}{nama wilayah}
  \item{\code{Luas}}{luas taman (m2)}
  \item{\code{Jumlah}}{jumlah taman}
}}
\source{
\url{http://data.bandung.go.id/dataset/luas-taman-kota-di-wilayah-kota-bandung}
}
\usage{
taman
}
\description{
Dataset ini berisi data luas taman kota di Sub Wilayah (SWK) Kota Bandung.
}
\keyword{datasets}
