% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{tanam_pohon1}
\alias{tanam_pohon1}
\title{Penanaman Pohon Di Kota Bandung 1}
\format{Tibble dengan 30 baris dan 3 variabel:
\describe{
  \item{\code{nama_kecamtan}}{nama kecamatan}
  \item{\code{pohon_pelindung}}{jumlah pohon pelindung yang ditanam}
  \item{\code{pohon_produktif}}{jumlah pohon produktif yang ditanam}
}}
\source{
\url{http://data.bandung.go.id/dataset/penanaman-pohon-di-kota-bandung/resource/76483bff-c81e-4381-a964-ced603e61d3e}
}
\usage{
tanam_pohon1
}
\description{
Dataset mengenai Rekapitulasi Penanaman Pohon Di Kota Bandung berdasarkan kecamatan pada tahun 2012.
}
\keyword{datasets}
