% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{tanam_pohon2}
\alias{tanam_pohon2}
\title{Penanaman Pohon Di Kota Bandung 2}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{\code{skpd}}{satuan kerja perangkat daerah}
  \item{\code{pohon_pelindung}}{jumlah pohon pelindung yang ditanam}
  \item{\code{pohon_produktif}}{jumlah pohon produktif yang ditanam}
}}
\source{
\url{http://data.bandung.go.id/dataset/penanaman-pohon-di-kota-bandung/resource/ae7da370-de47-4d9a-8d7f-1c6922fdaecb}
}
\usage{
tanam_pohon2
}
\description{
Dataset mengenai Rekapitulasi Penanaman Pohon Di Kota Bandung berdasarkan SKPD pada tahun 2013.
}
\keyword{datasets}
