% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{tps}
\alias{tps}
\title{Lokasi Tempat Pembuangan Sampah Kota Bandung}
\format{Tibble dengan 159 baris dan 11 variabel:
\describe{
  \item{\code{No}}{ }
  \item{\code{Wilayah}}{nama wilayah layanan}
  \item{\code{Kecamatan}}{nama kecamatan}
  \item{\code{TPS}}{nama tps}
  \item{\code{Lokasi}}{lokasi tps}
  \item{\code{Sampah_masuk}}{volume sampah masuk (m3/hari)}
  \item{\code{Sampah_diangkut}}{volume sampah diangkut ke tpa (m3/hari)}
  \item{\code{Ritasi}}{jumlah ritasi}
  \item{\code{Peruntukkan}}{peruntukan tps}
  \item{\code{Jadwal_pengangkutan}}{jadwal pengangkutan}
  \item{\code{Sumber_sampah}}{sumber sampah yang masuk ke tps}
}}
\source{
\url{http://data.bandung.go.id/dataset/lokasi-tempat-pembuangan-sampah-kota-bandung}
}
\usage{
tps
}
\description{
Dataset ini berisi data tempat pembuangan sampah kota Bandung.
}
\keyword{datasets}
