% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATASET.R
\docType{data}
\name{udara}
\alias{udara}
\title{Keadaan Udara Menurut Bulan di Kota Bandung}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 60 rows and 5 columns.}
\source{
\url{http://data.bandung.go.id/dataset/keadaan-udara-menurut-bulan-di-kota-bandung-2015/resource/67ef3f2c-aed8-4a77-bacf-94ea2b4e4248}
}
\usage{
udara
}
\description{
Dataset ini berisi data Keadaan Udara Menurut Bulan di Kota Bandung.
}
\keyword{datasets}
